EMP-583 Moved format functions to crmFieldFormatter
