+++ server/utils/crmFieldFormatter.ts
@@ -5,6 +5,10 @@ const currencyFormat = new Intl.NumberFormat('en-GB', {
   currency: 'GBP',
 })
 
+export const formatBooleanToYesNo = (value: boolean): string => {
+  return value ? 'Yes' : 'No'
+}
+
 export const formatCurrency = (value: string): string => {
   const number = Number(value)
   if (Number.isNaN(number)) {
@@ -23,6 +27,10 @@ export const formatDate = (value: string, dateFormat?: string): string => {
   }
 
   return format(date, 'd MMMM yyyy')
+}
+
+export const formatMultiline = (value: string) => {
+  return value.replace(/(\n|#13;\n|#13;)/g, '<br>')
 }
 
 export const formatTime = (value: string, dateFormat?: string): string => {
