+++ server/utils/utils.ts
@@ -39,14 +39,6 @@ export const currentIsoDate = (): string => {
   return format(new Date(), 'yyyy-MM-dd')
 }
 
-export const formatBooleanToYesNo = (value: boolean): string => {
-  return value ? 'Yes' : 'No'
-}
-
-export const formatMultiline = (value: string) => {
-  return value.replace(/(\n|#13;\n|#13;)/g, '<br>')
-}
-
 export const isNotEmpty = (value: string): boolean => {
   return value !== undefined && value !== null && value !== ''
 }
